backend/requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Login==0.5.0
flask-sqlacodegen==1.1.7
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==1.6.1
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
urllib3==1.25.10
uWSGI==2.0.19.1
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
scipy
pandas
datetime
matplotlib
numpy
openpyxl
seaborn
python-dateutil
requests
plotly
lightgbm
sklearn
apscheduler